Added make uninstall, refining outputs
